refactor: Post/BasicMarkdownPost -> MarkdownPost
