+++ src/pages/works/[id].tsx
@@ -1,4 +1,4 @@
-import { BasicMarkdownPost, BasicMarkdownPostProps } from "@/components/Post";
+import { MarkdownPost, MarkdownPostProps } from "@/components/MarkdownPost";
 import { ContentDirectoryNames, getContentIDs } from "@/modules/fs";
 import { getStaticMarkdoc } from "@/modules/markdown/server";
 import { GetStaticPaths, GetStaticProps } from "next";
@@ -8,7 +8,7 @@ interface Params {
 	id: string;
 }
 
-export default BasicMarkdownPost;
+export default MarkdownPost;
 
 export const getStaticPaths: GetStaticPaths = async () => {
 	const ids = await getContentIDs(ContentDirectoryNames.PROJECTS);
@@ -20,7 +20,7 @@ export const getStaticPaths: GetStaticPaths = async () => {
 	};
 };
 
-export const getStaticProps: GetStaticProps<BasicMarkdownPostProps, Params> = async (context) => {
+export const getStaticProps: GetStaticProps<MarkdownPostProps, Params> = async (context) => {
 	// This will never be thrown but it's here to satisfy TypeScript.
 	if (!context.params) throw "Error: works/[id]::getStaticProps context#params is falsy";
 	const { id: pageID } = context.params;
